+++ .env
@@ -1,4 +1,4 @@
 DB_USER=health-care
 DB_PASS=JbHLavveRkOeczgb
-# USER_NAME=sadiasalam6363
-# USER_PASS=RWiUMd8ob44VnmBB
+NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY="pk_test_51NFsUALb1KL8Ug5lO28Ay7gsntSQgD4QCUVZAo1hZPRNPXhxNsP4AnEn4R5ZxAaM045xsdKf3XoMRV39s0273okB00XSqUPwgj"
+STRIPE_SECRET_KEY="sk_test_51NFsUALb1KL8Ug5ldeNmJavolEWnwPEbtfaQv7yRj5MF6gLtAvoEr3J9LtCCEIL7sc2sRilsjFTvqPgkA36abm4100Oj2JzQSY"
