+++ src/Components/Footer/Footer.jsx
@@ -3,94 +3,66 @@ import { FaYoutube, FaFacebookF, FaTwitter } from "react-icons/fa";
 
 const Footer = () => {
   return (
-    <>
-    <footer className="footer p-10 bg-black text-white">
-  <nav>
-    <header className="footer-title">Services</header> 
-    <a className="link link-hover">Branding</a> 
-    <a className="link link-hover">Design</a> 
-    <a className="link link-hover">Marketing</a> 
-    <a className="link link-hover">Advertisement</a>
-  </nav> 
-  <nav>
-    <header className="footer-title">Company</header> 
-    <a className="link link-hover">About us</a> 
-    <a className="link link-hover">Contact</a> 
-    <a className="link link-hover">Jobs</a> 
-    <a className="link link-hover">Press kit</a>
-  </nav> 
-  <nav>
-    <header className="footer-title">Legal</header> 
-    <a className="link link-hover">Terms of use</a> 
-    <a className="link link-hover">Privacy policy</a> 
-    <a className="link link-hover">Cookie policy</a>
-  </nav>
-</footer> 
-<footer className="footer px-10 py-4 border-t bg-base-200 text-base-content border-base-300">
-  <aside className="items-center grid-flow-col">
-    <svg width="24" height="24" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" className="fill-current"><path d="M22.672 15.226l-2.432.811.841 2.515c.33 1.019-.209 2.127-1.23 2.456-1.15.325-2.148-.321-2.463-1.226l-.84-2.518-5.013 1.677.84 2.517c.391 1.203-.434 2.542-1.831 2.542-.88 0-1.601-.564-1.86-1.314l-.842-2.516-2.431.809c-1.135.328-2.145-.317-2.463-1.229-.329-1.018.211-2.127 1.231-2.456l2.432-.809-1.621-4.823-2.432.808c-1.355.384-2.558-.59-2.558-1.839 0-.817.509-1.582 1.327-1.846l2.433-.809-.842-2.515c-.33-1.02.211-2.129 1.232-2.458 1.02-.329 2.13.209 2.461 1.229l.842 2.515 5.011-1.677-.839-2.517c-.403-1.238.484-2.553 1.843-2.553.819 0 1.585.509 1.85 1.326l.841 2.517 2.431-.81c1.02-.33 2.131.211 2.461 1.229.332 1.018-.21 2.126-1.23 2.456l-2.433.809 1.622 4.823 2.433-.809c1.242-.401 2.557.484 2.557 1.838 0 .819-.51 1.583-1.328 1.847m-8.992-6.428l-5.01 1.675 1.619 4.828 5.011-1.674-1.62-4.829z"></path></svg>
-    <p>ACME Industries Ltd. <br/>Providing reliable tech since 1992</p>
-  </aside> 
-  <nav className="md:place-self-center md:justify-self-end">
-    <div className="grid grid-flow-col gap-4">
-    <FaTwitter style={{ fontSize: "2rem" }}></FaTwitter>
-            <FaYoutube style={{ fontSize: "2rem" }}></FaYoutube>
-            <FaFacebookF style={{ fontSize: "2rem" }}></FaFacebookF>
-    </div>
-  </nav>
-</footer>
-    </>
-  );
-};
-
-export default Footer;
-{/* <div>
-      <footer className="footer p-16 bg-[#262b35] grid md:grid-cols-4 grid-cols-1  text-[#8890a4]">
-        <div>
+    <div>
+     
+      <footer className="footer p-10 bg-black text-white">
+        <nav>
+          <header className="footer-title">Services</header>
+          <a className="link link-hover">Branding</a>
+          <a className="link link-hover">Design</a>
+          <a className="link link-hover">Marketing</a>
+          <a className="link link-hover">Advertisement</a>
+        </nav>
+        <nav>
+          <header className="footer-title">Company</header>
+          <a className="link link-hover">About us</a>
+          <a className="link link-hover">Contact</a>
+          <a className="link link-hover">Jobs</a>
+          <a className="link link-hover">Press kit</a>
+        </nav>
+        <nav>
+          <header className="footer-title">Legal</header>
+          <a className="link link-hover">Terms of use</a>
+          <a className="link link-hover">Privacy policy</a>
+          <a className="link link-hover">Cookie policy</a>
+        </nav>
+      </footer>
+      <footer className="footer px-10 py-4 border-t bg-black text-white border-base-300">
+        <aside className="items-center grid-flow-col">
           <h1 className="flex items-center text-2xl font-bold">
             <span>
               {" "}
               <img
                 src="https://i.ibb.co/NpptRzh/download-removebg-preview.png"
                 alt=""
-                className="h-10 w-10"
+                className="h-20 w-20"
               />
             </span>
-            <span>HealthCare</span>
+            <span>
+              HealthCare
+              <br />
+              <span className="text-sm text-slate-400 font-normal ">
+                {" "}
+                providing services Since 2015
+              </span>
+            </span>
           </h1>
-
-          <p className="font-bold">
-            <br />
-            Providing reliable treatment since 2000
-          </p>
-        </div>
-
-        <div className="mx-auto text-center">
-          <span className="  font-bold">Navigation's</span>
-          <Link href="/">Home</Link>
-          <Link href="/about">About</Link>
-          <Link href="/contacts">Contacts</Link>
-          <Link href="/doctors">Doctors</Link>
-        </div>
-        <div className="mx-auto">
-          <span className="font-bold">Contact</span>
-          <div>
-            <p className="text-md">
-              <strong>Email</strong> : jihanstk@gmail.com
-            </p>
-            <p className="text-md">
-              <strong>Address</strong> : Sector-12,Uttara Dhaka
-            </p>
+        </aside>
+        <nav className="md:place-self-center md:justify-self-end">
+          <div className="grid grid-flow-col gap-4">
+            <div className="mx-auto">
+              <span className="font-bold text-center ">Social</span>
+              <div className="grid grid-flow-col gap-4 text-white">
+                <FaTwitter style={{ fontSize: "2rem" }}></FaTwitter>
+                <FaYoutube style={{ fontSize: "2rem" }}></FaYoutube>
+                <FaFacebookF style={{ fontSize: "2rem" }}></FaFacebookF>
+              </div>
+            </div>
           </div>
-        </div>
-        <div className="mx-auto">
-          <span className="font-bold text-center ">Social</span>
-          <div className="grid grid-flow-col gap-4 text-white">
-          
-          </div>
-        </div>
-      </footer> */}
-      {/* <div className="footer footer-center font-bold bg-[#262b35] text-[#8890a4]  p-4">
-        <p>Copyright © 2023 - All right reserved by HealthCare Ltd</p>
-      </div>
-    </div> */}
+        </nav>
+      </footer>
+    </div>
+  );
+};
+
+export default Footer;
