+++ src/app/(with-navbar)/page.js
@@ -1,4 +1,6 @@
-
+// import About from "@/Components/About/About";
+// import AllDoctors from "@/Components/AllDoctors/AllDoctors";
+// import About from "@/Components/About/About";
 import Appointment from "@/Components/Appointment/Appointment";
 import BannerSlider from "@/Components/Banner/BannerSlider";
 import OurDoctors from "@/Components/Content/OurDoctors/OurDoctors";
@@ -10,7 +12,7 @@ export default function Home() {
   return (
     <main>
       <BannerSlider></BannerSlider>
-     
+      {/* <About></About> */}
       <OurDoctors />
       <OurTest />
       <PopularDoctors></PopularDoctors>
