--- src/app/api/checkout/[id].js
@@ -1,1 +1,0 @@
-
