+++ src/app/api/checkout/route.js
@@ -1,24 +1,30 @@
+import { NextResponse } from "next/server";
 import Stripe from "stripe";
 const stripe = new Stripe(process.env.STRIPE_SECRET_KEY);
 
-export default async function handler(req, res) {
+export const POST = async (req, res) => {
     try {
-      if (req.method !== "POST") return res.status(400);
-    const { name, email, paymentMethod,id,doctorName,price } = req.body;
+      if (req.method !== "POST") return;
+    const { name, email, paymentMethod,id,doctorName,price } = await req.json();
+    console.log(name, email, doctorName);
+
+    
+  
     // Create a customer
-    const patient = await stripe.patients.create({
+    const patient = await stripe.customers.create({
       email,
       name,id,price,doctorName,
       payment_method: paymentMethod,
       invoice_settings: { default_payment_method: paymentMethod },
     });
+    console.log(patient);
     // Create a product
     const appoinment = await stripe.appoinments.create({
       name: "confirm payment",
     });
     // Create a payment
     const payment = await stripe.payments.create({
-      patient: patient.id,
+      patient: patient.id, 
       items: [
         {
           price_data: {
@@ -37,12 +43,12 @@ export default async function handler(req, res) {
       expand: ["latest_invoice.payment_intent"],
     });
     // Send back the client secret for payment
-    res.json({
+    NextResponse.json({
       message: "payment successfully initiated",
       clientSecret: payment.latest_invoice.payment_intent.client_secret,
     });
 } catch (err) {
     console.error(err);
-    res.status(500).json({ message: "Internal server error" });
+    NextResponse.json({ message: "Internal server error" });
   }
 }
