--- src/app/api/route.js
@@ -1,4 +1,0 @@
-import { NextResponse } from "next/server";
-export const GET = async () => {
-  return NextResponse.json({ message: "hello" });
-};
